requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.6
boto3==1.9.164
botocore==1.12.164
cffi==1.12.3
cryptography==2.7
docutils==0.14
importlib-metadata==0.17
jmespath==0.9.4
more-itertools==7.0.0
packaging==19.0
paramiko==2.4.2
pdfminer3k==1.3.1
pluggy==0.12.0
ply==3.11
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.4.0
pytest==4.6.2
python-dateutil==2.8.0
s3transfer==0.2.1
scp==0.13.2
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.1
